Introduced consumer locator. First step to abstract runtime dependency on windsor.
