+++ Rhino.ServiceBus/Impl/DefaultServiceBus.cs
@@ -3,20 +3,15 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
 using Castle.MicroKernel;
-using Castle.MicroKernel.Context;
-using Castle.MicroKernel.Proxy;
 using log4net;
 using Rhino.ServiceBus.Exceptions;
 using Rhino.ServiceBus.Internal;
 using Rhino.ServiceBus.MessageModules;
 using Rhino.ServiceBus.Messages;
-using Rhino.ServiceBus.Sagas;
 
 namespace Rhino.ServiceBus.Impl
 {
-	using Queues.Utils;
-
-	public class DefaultServiceBus : IStartableServiceBus
+    public class DefaultServiceBus : IStartableServiceBus
     {
         private readonly IKernel kernel;
 
@@ -28,18 +23,21 @@ namespace Rhino.ServiceBus.Impl
         private readonly MessageOwnersSelector messageOwners;
     	[ThreadStatic] public static object currentMessage;
         private readonly IEndpointRouter endpointRouter;
+	    private readonly IConsumerLocator consumerLocator;
 
-        public DefaultServiceBus(
+	    public DefaultServiceBus(
             IKernel kernel,
             ITransport transport,
             ISubscriptionStorage subscriptionStorage,
             IReflection reflection,
             IMessageModule[] modules,
             MessageOwner[] messageOwners, 
-            IEndpointRouter endpointRouter)
+            IEndpointRouter endpointRouter,
+            IConsumerLocator consumerLocator)
         {
             this.transport = transport;
             this.endpointRouter = endpointRouter;
+            this.consumerLocator = consumerLocator;
             this.messageOwners = new MessageOwnersSelector(messageOwners, endpointRouter);
             this.subscriptionStorage = subscriptionStorage;
             this.reflection = reflection;
@@ -47,13 +45,10 @@ namespace Rhino.ServiceBus.Impl
             this.kernel = kernel;
         }
 
-		
         public IMessageModule[] Modules
         {
             get { return modules; }
         }
-
-        #region IStartableServiceBus Members
 
         public event Action<Reroute> ReroutedEndpoint;
 
@@ -343,8 +338,6 @@ namespace Rhino.ServiceBus.Impl
             }
         }
 
-        #endregion
-
         private bool PublishInternal(object[] messages)
         {
             if (messages == null)
@@ -372,7 +365,7 @@ namespace Rhino.ServiceBus.Impl
 
         public bool Transport_OnMessageArrived(CurrentMessageInformation msg)
         {
-            object[] consumers = GatherConsumers(msg);
+            var consumers = consumerLocator.GatherConsumers(msg.Message);
         	
 			if (consumers.Length == 0)
             {
@@ -443,143 +436,10 @@ namespace Rhino.ServiceBus.Impl
                 reflection.InvokeSagaPersisterSave(persister, saga);
         }
 
+        [Obsolete("Use IConsumerLocator.GatherConsumers instead.")]
         public object[] GatherConsumers(CurrentMessageInformation msg)
         {
-            object[] sagas = GetSagasFor(msg.Message);
-            var sagaMessage = msg.Message as ISagaMessage;
-
-            var msgType = msg.Message.GetType();
-            object[] instanceConsumers = subscriptionStorage
-                .GetInstanceSubscriptions(msgType);
-
-            var consumerTypes = reflection.GetGenericTypesOfWithBaseTypes(typeof(ConsumerOf<>), msg.Message);
-            var occasionalConsumerTypes = reflection.GetGenericTypesOfWithBaseTypes(typeof(OccasionalConsumerOf<>), msg.Message);
-            var consumers = GetAllNonOccasionalConsumers(consumerTypes, occasionalConsumerTypes, sagas);
-            for (var i = 0; i < consumers.Length; i++)
-            {
-                var saga = consumers[i] as IAccessibleSaga;
-                if (saga == null)
-                    continue;
-
-                // if there is an existing saga, we skip the new one
-                var type = saga.GetType();
-                if (sagas.Any(type.IsInstanceOfType))
-                {
-                    kernel.ReleaseComponent(consumers[i]);
-                    consumers[i] = null;
-                    continue;
-                }
-                // we do not create new sagas if the saga is not initiated by
-                // the message
-                var initiatedBy = reflection.GetGenericTypeOf(typeof(InitiatedBy<>),msgType);
-                if(initiatedBy.IsInstanceOfType(saga)==false)
-                {
-                    kernel.ReleaseComponent(consumers[i]);
-                    consumers[i] = null;
-                    continue;
-                }
-
-                saga.Id = sagaMessage != null ?
-                    sagaMessage.CorrelationId :
-                    GuidCombGenerator.Generate();
-            }
-            return instanceConsumers
-                .Union(sagas)
-                .Union(consumers.Where(x => x != null))
-                .ToArray();
-        }
-
-        /// <summary>
-        /// Here we don't use ResolveAll from Windsor because we want to get an error
-        /// if a component exists which isn't valid
-        /// </summary>
-        private object[] GetAllNonOccasionalConsumers(ICollection<Type> consumerTypes, ICollection<Type> occasionalConsumerTypes, IEnumerable<object> instanceOfTypesToSkipResolving)
-        {
-            var allHandlers = new List<IHandler>();
-            foreach (var consumerType in consumerTypes)
-            {
-                var handlers = kernel.GetAssignableHandlers(consumerType);
-                allHandlers.AddRange(handlers);
-            }
-
-            var consumers = new List<object>(allHandlers.Count);
-            foreach (var handler in allHandlers)
-            {
-                var implementation = handler.ComponentModel.Implementation;
-                var occasionalConsumerFound = false;
-                foreach (var occasionalConsumerType in occasionalConsumerTypes)
-                {
-                    if (occasionalConsumerType.IsAssignableFrom(implementation))
-                    {
-                        occasionalConsumerFound = true;
-                        break;
-                    }
-                }
-                if (occasionalConsumerFound)
-                    continue;
-                
-                if (instanceOfTypesToSkipResolving.Any(x => x.GetType() == implementation))
-                    continue;
-
-                consumers.Add(handler.Resolve(CreationContext.Empty));
-            }
-            return consumers.ToArray();
-        }
-
-        private object[] GetSagasFor(object message)
-        {
-            var instances = new List<object>();
-
-            Type orchestratesType = reflection.GetGenericTypeOf(typeof(Orchestrates<>), message);
-            Type initiatedByType = reflection.GetGenericTypeOf(typeof(InitiatedBy<>), message);
-
-            var handlers = kernel.GetAssignableHandlers(orchestratesType)
-                                                   .Union(kernel.GetAssignableHandlers(initiatedByType));
-
-            foreach (IHandler sagaHandler in handlers)
-            {
-                Type sagaType = sagaHandler.ComponentModel.Implementation;
-                
-                //first try to execute any saga finders.
-                Type sagaFinderType = reflection.GetGenericTypeOf(typeof (ISagaFinder<,>), sagaType, ProxyUtil.GetUnproxiedType(message));
-                var sagaFinderHandlers = kernel.GetAssignableHandlers(sagaFinderType);
-                foreach (var sagaFinderHandler in sagaFinderHandlers)
-                {
-                    try
-                    {
-                        var sagaFinder = kernel.Resolve(sagaFinderHandler.Service);
-                        var saga = reflection.InvokeSagaFinderFindBy(sagaFinder, message);
-                        if (saga != null)
-                            instances.Add(saga);
-                    }
-                    finally
-                    {
-                        kernel.ReleaseComponent(sagaFinderHandler);
-                    }
-                }
-
-                //we will try to use an ISagaMessage's Correlation id next.
-                var sagaMessage = message as ISagaMessage;
-                if (sagaMessage == null)
-                    continue;
-
-                Type sagaPersisterType = reflection.GetGenericTypeOf(typeof(ISagaPersister<>),
-                                                                     sagaType);
-
-                object sagaPersister = kernel.Resolve(sagaPersisterType);
-                try
-                {
-                    object sagas = reflection.InvokeSagaPersisterGet(sagaPersister, sagaMessage.CorrelationId);
-                    if (sagas == null)
-                        continue;
-                    instances.Add(sagas);
-                }
-                finally
-                {
-                    kernel.ReleaseComponent(sagaPersister);
-                }
-            }
-            return instances.ToArray();
+            return consumerLocator.GatherConsumers(msg.Message);
         }
     }
 }
