+++ Rhino.ServiceBus/Internal/IConsumerLocator.cs
@@ -1,0 +1,7 @@
+namespace Rhino.ServiceBus.Internal
+{
+    public interface IConsumerLocator
+    {
+        object[] GatherConsumers(object message);
+    }
+}
