+++ Rhino.ServiceBus/Rhino.ServiceBus.csproj
@@ -94,8 +94,10 @@
     <Compile Include="Hosting\RemoteAppDomainLoadBalancerHost.cs" />
     <Compile Include="Hosting\RhinoQueuesHostConfiguration.cs" />
     <Compile Include="ICustomizeMessageHeaders.cs" />
+    <Compile Include="Impl\CastleConsumerLocator.cs" />
     <Compile Include="Impl\EndpointRouter.cs" />
     <Compile Include="Endpoint.cs" />
+    <Compile Include="Internal\IConsumerLocator.cs" />
     <Compile Include="Internal\IEndpointRouter.cs" />
     <Compile Include="IServiceBusAware.cs" />
     <Compile Include="Impl\DefaultServiceBus.cs" />
